Hotfix/merge deployv1 to develop

+++ src/main/java/com/EIPplatform/model/dto/report/report05/ReportA05DTO.java
@@ -1,6 +1,5 @@
 package com.EIPplatform.model.dto.report.report05;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -31,7 +30,7 @@ public class ReportA05DTO {
     String reportingPeriod;
     String reviewNotes;
     String inspectionRemedyReport;
-    BigDecimal completionPercentage;
+    Double completionPercentage;
     LocalDateTime createdAt;
 
     WasteWaterDataDTO wasteWaterData;
